Add common debug wrapper
WIN32, AVR and POSIX ports each use different debug print facilities.
Create a common wrapper API so that generic code can emit debug output
on any platform.

This patch does not modify the platform-specific code to use the new
macro because the call-sites are in code that needs to be consolidated
into the generic code base anyway. A subsequent patch will fix this.

Signed-off-by: Grant Likely <grant.likely@secretlab.ca>