Consolidate _VISCA_send_packet()
Each platform had a different implementation of _VISCA_send_packet(),
even though the implementations are logically identical. The only
difference is the method of debug output and how the send function is
called.

Consolidate all the implementations into the common code and use the new
_VISCA_debug() macro for sending output. For the win32 version, replace
the inline code for sending data with a call to
_VISCA_send_packet_data()

Signed-off-by: Grant Likely <grant.likely@secretlab.ca>