Create common core pan/tilt function
All of the pan/tilt direction functions are nearly identical. Refactor
with a new common core function 'VISCA_set_pantilt()' that accepts
signed speed values. This allows users of the library to call a single
function regardless of the direction, where negative values represent
panning left or tilting down. Refactoring in this way also reduces the
size of the library.

Signed-off-by: Grant Likely <grant.likely@secretlab.ca>