win32: check error codes on win32 library calls
The win32 code wasn't checking any return values. Restructure the code
to check for failures when setting up the UART. Have chosen to use the
Linux-style 'goto err' exception handling approach to collect unwinding
from the error at the bottom of the function. This avoids both deep
nesting of if/else blocks and duplicate unwind code in each failure
condition.

Signed-off-by: Grant Likely <grant.likely@secretlab.ca>