Fix UART configuration to return data immediately
Change the UART configuration to return any received bytes immediately,
or timeout and return an error if nothing is received. Processing the
bytes immediately will be important for a future patch that enables
non-blocking operation. Adding the timeout behaviour keeps the test
programs from hanging if a camera is not connected to the UART.

Signed-off-by: Grant Likely <grant.likely@secretlab.ca>