Use VISCA_INPUT_BUFFER_SIZE for output buffers too
The VISCA input buffer size is defined by VISCA_INPUT_BUFFER_SIZE, but
the output buffer is arbitrarily 32 bytes. Rework the code to use the
preprocessor #define for the input buffer. This also reduces the input
buffer from 1024 to 32. This is fine because by the spec the largest
VISCA packet is 16 bytes. 1024 really isn't needed.

Signed-off-by: Grant Likely <grant.likely@secretlab.ca>