Drop address property from VISCAInterface_t
In VISCA, the controller address is always 0, so there is no need to
track it in a member variable. Theoretically this library could be used
to implement a VISCA device, not just controller, but changes would be
needed first. For instance, the library doesn't know how to respond to
VISCA commands or inquiries. Remove the member variable for now as it
simplifies the code. If can be added back later if needed.

Signed-off-by: Grant Likely <grant.likely@secretlab.ca>