Create asynchronous receive state machine
Restructure receive logic to be reentrant to make it possible for
commands to be executed asynchronously. With this patch a new
VISCA_loop() function is created that polls the serial port for new
packets. When bytes are received, they are piped into the receive packet
and then the whole packet is parsed when the end of frame 0xff
terminator is received.

If the command is an inquiry that returns data, then the reply packet is
copied into the caller's buffer while in the packet processing function
so that the caller can process it later.

The existing command helpers still busywait for the command to
complete. A future patch will rework them to allow asynchronous calls.

Signed-off-by: Grant Likely <grant.likely@secretlab.ca>